chore: move locate-date.ts to lib
